datacsm-kernel: qcom-aw-phy: Reset the equalization mode during teardown
Changes to reset the equalization mode which was set with the
previous lane bring up attempt, so that it does not impact
the next lane bring up attempt. This is to cover the scenario
where lane bring up was done in ANLT mode and then the switch
is made to do lane bring up in manual EQ mode.

Change-Id: I5512a5c736ad2d1301b4b14e5b7fb11908ab80d2