datacsm-kernel: ecpri_core: Enabled DL/UL route to DMA for egress logging on RU/X100
For FH logging to work for C and U plane data, we need to make sure
DL and UL packets are routed to DMA on RU/X100 independent of the logging direction the user
has selected. This patch removes the direction specific check and adds
routes for both the directions.

Change-Id: I20095de1245c24e82f512760cb47fcea85e99791